backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psutil==5.9.8
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
python-dotenv==1.0.0
GPUtil==1.4.0
django-prometheus==2.3.1
